+++ lib/proxy_fetcher.rb
@@ -46,7 +46,6 @@ module ProxyFetcher
 
   # Main ProxyFetcher module.
   class << self
-
     ##
     # Returns ProxyFetcher configuration.
     #
@@ -57,9 +56,9 @@ module ProxyFetcher
     #   ProxyFetcher.config
     #
     #   #=> #<ProxyFetcher::Configuration:0x0000000241eec8 @user_agent="Mozilla/5.0, ...", @pool_size=10,
-    #           @client_timeout=3, @proxy_validation_timeout=3, @provider_proxies_load_timeout=30,
-    #           @http_client=ProxyFetcher::HTTPClient, @proxy_validator=ProxyFetcher::ProxyValidator,
-    #           @providers=[:free_proxy_list, ...], @adapter=ProxyFetcher::Document::NokogiriAdapter>
+    #          @client_timeout=3, @proxy_validation_timeout=3, @provider_proxies_load_timeout=30,
+    #          @http_client=ProxyFetcher::HTTPClient, @proxy_validator=ProxyFetcher::ProxyValidator,
+    #          @providers=[:free_proxy_list, ...], @adapter=ProxyFetcher::Document::NokogiriAdapter>
     #
     def config
       @__config_definition_lock__.synchronize do